add sign in with email and pass with firebase and creat home screen
