--- lib/login.dart
@@ -1,98 +1,0 @@
-import 'package:flutter/material.dart';
-
-// ignore: must_be_immutable
-class LoginScreen extends StatefulWidget {
-  LoginScreen({Key? key}) : super(key: key);
-
-  @override
-  State<LoginScreen> createState() => _LoginScreenState();
-  bool isActive = true;
-}
-
-class _LoginScreenState extends State<LoginScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      resizeToAvoidBottomInset: false,
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Center(child: Image.asset(height: 150, 'assets/images/login.png')),
-          const Text(
-            'Login Now',
-            style: TextStyle(fontSize: 25),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: TextFormField(
-              decoration: const InputDecoration(
-                  prefixIcon: Icon(Icons.email_outlined),
-                  hintText: 'Enter Your Email',
-                  label: Text('Email'),
-                  focusedBorder: OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.green),
-                  ),
-                  enabledBorder: OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.white),
-                  )),
-            ),
-          ),
-          const SizedBox(
-            height: 30,
-          ),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: TextFormField(
-              obscureText: widget.isActive,
-              decoration: InputDecoration(
-                  suffixIcon: InkWell(
-                    onTap: togglePassword,
-                    child: widget.isActive
-                        ? const Icon(Icons.visibility_off)
-                        : const Icon(Icons.visibility),
-                  ),
-                  prefixIcon: const Icon(Icons.lock_open_outlined),
-                  hintText: 'Enter Your Password',
-                  label: const Text('password'),
-                  focusedBorder: const OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.green),
-                  ),
-                  enabledBorder: const OutlineInputBorder(
-                    borderSide: BorderSide(color: Colors.white),
-                  )),
-            ),
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              const Text('Are You Have Account ? '),
-              TextButton(
-                onPressed: () {
-                  Navigator.pushNamed(context, 'Register');
-                },
-                child: const Text(
-                  'Register Now',
-                ),
-              ),
-            ],
-          ),
-          ElevatedButton(
-              style: const ButtonStyle(
-                  backgroundColor: MaterialStatePropertyAll(Colors.deepPurple),
-                  elevation: MaterialStatePropertyAll(10),
-                  textStyle: MaterialStatePropertyAll(
-                      TextStyle(fontSize: 20, fontWeight: FontWeight.bold)),
-                  padding: MaterialStatePropertyAll(
-                      EdgeInsets.symmetric(horizontal: 20, vertical: 40))),
-              onPressed: () {},
-              child: const Text('Login'))
-        ],
-      ),
-    );
-  }
-
-  void togglePassword() {
-    widget.isActive = !widget.isActive;
-    setState(() {});
-  }
-}
